Refactor map fast search into class
