+++ .gitignore
@@ -1,6 +1,6 @@
 /client/node_modules/
 /server/node_modules/
 node_modules/
-package-lock.json
+*/package-lock.json
 /server/.map-cache/
 /server/grid.json
