+++ server/src/services/csv-reader.service.ts
@@ -2,11 +2,8 @@ import { Injectable, RequestTimeoutException } from '@nestjs/common';
 import * as fs from 'fs';
 import csv = require('csv-parser');
 
-
 @Injectable()
-export class csvReader {
-    constructor() { }
-
+export class CsvReader {
     readCSV(csvfilename: string, mapper: (row) => void): void {
         fs.createReadStream(csvfilename)
             .pipe(csv())
